modify price
minor fixes to the price section and changed the position of the button
